Rename the provider tests.
