+++ src/test/java/org/jboss/logging/JBossLogManagerProviderTestCase.java
@@ -20,10 +20,14 @@ package org.jboss.logging;
 
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.logging.Handler;
 import java.util.logging.Level;
 import java.util.logging.LogRecord;
 
+import org.jboss.logmanager.LogContext;
+import org.jboss.logmanager.LogContextSelector;
+import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
@@ -33,50 +37,75 @@ import org.junit.jupiter.api.Test;
 /**
  * @author <a href="mailto:jperkins@redhat.com">James R. Perkins</a>
  */
-public class JulManagerTestCase extends AbstractLoggerTestCase {
+public class JBossLogManagerProviderTestCase extends AbstractLoggerTestCase {
+    private static final LogContextSelector DEFAULT_SELECTOR = LogContext.getLogContextSelector();
+    private static final AtomicBoolean SET_LOG_MANAGER = new AtomicBoolean(true);
+
+    private LogContext logContext;
     private TestHandler handler;
     private Logger logger;
 
     @BeforeAll
     public static void setup() {
-        System.setProperty("org.jboss.logging.provider", "jdk");
+        SET_LOG_MANAGER.set(System.getProperty("java.util.logging.manager") == null);
+        if (SET_LOG_MANAGER.get()) {
+            System.setProperty("java.util.logging.manager", "org.jboss.logmanager.LogManager");
+        }
+        System.setProperty("org.jboss.logging.provider", "jboss");
+    }
+
+    @AfterAll
+    public static void tearDown() {
+        if (SET_LOG_MANAGER.get()) {
+            System.clearProperty("java.util.logging.manager");
+        }
     }
 
     @BeforeEach
     public void setupLogContext() {
+        logContext = LogContext.create();
+        LogContext.setLogContextSelector(() -> logContext);
         logger = Logger.getLogger(getClass());
         handler = createHandler(logger.getName());
     }
 
     @AfterEach
-    public void removeHandler() {
-        java.util.logging.Logger.getLogger(logger.getName()).removeHandler(handler);
-        handler.close();
+    public void closeLogContext() throws Exception {
+        logContext.close();
+        LogContext.setLogContextSelector(DEFAULT_SELECTOR);
     }
 
     @Test
     public void testLogger() {
-        Assertions.assertTrue(logger instanceof JDKLogger);
+        Assertions.assertEquals(JBossLogManagerLogger.class, logger.getClass());
     }
 
     @Test
     public void testMdc() {
         MDC.put("test.key", "value");
         Assertions.assertEquals("value", MDC.get("test.key"));
+        Assertions.assertEquals("value", org.jboss.logmanager.MDC.get("test.key"));
     }
 
     @Test
     public void testNdc() {
         NDC.push("value1");
         NDC.push("value2");
-        Assertions.assertEquals("value2", NDC.peek());
-        Assertions.assertEquals("value1 value2", NDC.get());
+        final String expectedValue = "value1.value2";
+        Assertions.assertEquals(expectedValue, NDC.peek());
+        Assertions.assertEquals(expectedValue, NDC.get());
         Assertions.assertEquals(2, NDC.getDepth());
+
+        // Test the log manager values
+        Assertions.assertEquals(expectedValue, org.jboss.logmanager.NDC.get());
+        Assertions.assertEquals(2, org.jboss.logmanager.NDC.getDepth());
 
         // Pop the stack
         Assertions.assertEquals("value2", NDC.pop());
         Assertions.assertEquals(1, NDC.getDepth());
         Assertions.assertEquals("value1", NDC.get());
+        Assertions.assertEquals("value1", org.jboss.logmanager.NDC.get());
+        Assertions.assertEquals(1, org.jboss.logmanager.NDC.getDepth());
     }
 
     @Override
@@ -107,7 +136,7 @@ public class JulManagerTestCase extends AbstractLoggerTestCase {
 
     private TestHandler createHandler(final String loggerName) {
         final TestHandler handler = new TestHandler();
-        final java.util.logging.Logger julLogger = java.util.logging.Logger.getLogger(loggerName);
+        final java.util.logging.Logger julLogger = logContext.getLogger(loggerName);
         julLogger.addHandler(handler);
         julLogger.setLevel(Level.ALL);
         return handler;
