+++ src/test/java/org/jboss/logging/Log4jProviderTestCase.java
@@ -21,52 +21,49 @@ package org.jboss.logging;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
 
-import ch.qos.logback.classic.Level;
-import ch.qos.logback.classic.LoggerContext;
-import ch.qos.logback.classic.spi.ILoggingEvent;
-import ch.qos.logback.core.AppenderBase;
+import org.apache.log4j.AppenderSkeleton;
+import org.apache.log4j.Level;
+import org.apache.log4j.spi.LoggingEvent;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.slf4j.LoggerFactory;
 
 /**
  * @author <a href="mailto:jperkins@redhat.com">James R. Perkins</a>
  */
-public class Slf4jManagerTestCase extends AbstractLoggerTestCase {
+public class Log4jProviderTestCase extends AbstractLoggerTestCase {
     private TestAppender appender;
     private Logger logger;
 
     @BeforeAll
     public static void setup() {
-        System.setProperty("org.jboss.logging.provider", "slf4j");
+        System.setProperty("org.jboss.logging.provider", "log4j");
     }
 
     @BeforeEach
     public void setupLogContext() {
         logger = Logger.getLogger(getClass());
-        appender = createHandler(logger.getName());
+        appender = createAppender(logger.getName());
     }
 
     @AfterEach
     public void removeAppender() {
-        ch.qos.logback.classic.Logger lbLogger = (ch.qos.logback.classic.Logger) LoggerFactory.getLogger(logger.getName());
-        lbLogger.detachAppender(appender);
-        appender.stop();
+        org.apache.log4j.Logger.getLogger(logger.getName()).removeAppender(appender);
+        appender.close();
     }
 
     @Test
     public void testLogger() {
-        Assertions.assertTrue(logger instanceof Slf4jLogger || logger instanceof Slf4jLocationAwareLogger);
+        Assertions.assertTrue(logger instanceof Log4jLogger);
     }
 
     @Test
     public void testMdc() {
         MDC.put("test.key", "value");
         Assertions.assertEquals("value", MDC.get("test.key"));
-        Assertions.assertEquals("value", org.slf4j.MDC.get("test.key"));
+        Assertions.assertEquals("value", org.apache.log4j.MDC.get("test.key"));
     }
 
     @Test
@@ -77,10 +74,16 @@ public class Slf4jManagerTestCase extends AbstractLoggerTestCase {
         Assertions.assertEquals("value1 value2", NDC.get());
         Assertions.assertEquals(2, NDC.getDepth());
 
+        // Test the log manager values
+        Assertions.assertEquals("value1 value2", org.apache.log4j.NDC.get());
+        Assertions.assertEquals(2, org.apache.log4j.NDC.getDepth());
+
         // Pop the stack
         Assertions.assertEquals("value2", NDC.pop());
         Assertions.assertEquals(1, NDC.getDepth());
         Assertions.assertEquals("value1", NDC.get());
+        Assertions.assertEquals("value1", org.apache.log4j.NDC.get());
+        Assertions.assertEquals(1, org.apache.log4j.NDC.getDepth());
     }
 
     @Override
@@ -90,20 +93,18 @@ public class Slf4jManagerTestCase extends AbstractLoggerTestCase {
 
         Assertions.assertTrue(logger.isEnabled(level), String.format("Logger not enabled for level %s", level));
 
-        final ILoggingEvent event = appender.queue.poll();
+        final LoggingEvent event = appender.queue.poll();
         Assertions.assertNotNull(event, String.format("No record found for %s", level));
-        final Logger.Level translatedLevel = level == Logger.Level.FATAL ? Logger.Level.ERROR : level;
-        Assertions.assertEquals(translatedLevel.name(), event.getLevel().toString());
-        Assertions.assertEquals(msg, event.getFormattedMessage());
+        Assertions.assertEquals(level.name(), event.getLevel().toString());
+        Assertions.assertEquals(msg, event.getMessage());
     }
 
     @Override
     void testLog(final String msg, final Logger.Level level) {
-        final ILoggingEvent event = appender.queue.poll();
+        final LoggingEvent event = appender.queue.poll();
         Assertions.assertNotNull(event, String.format("No record found for %s", level));
-        final Logger.Level translatedLevel = level == Logger.Level.FATAL ? Logger.Level.ERROR : level;
-        Assertions.assertEquals(translatedLevel.name(), event.getLevel().toString());
-        Assertions.assertEquals(msg, event.getFormattedMessage());
+        Assertions.assertEquals(level.name(), event.getLevel().toString());
+        Assertions.assertEquals(msg, event.getMessage());
     }
 
     @Override
@@ -111,24 +112,30 @@ public class Slf4jManagerTestCase extends AbstractLoggerTestCase {
         return logger;
     }
 
-    private static TestAppender createHandler(final String loggerName) {
-        final LoggerContext context = (LoggerContext) LoggerFactory.getILoggerFactory();
+    private static TestAppender createAppender(final String loggerName) {
         final TestAppender appender = new TestAppender();
-        appender.setContext(context);
-        appender.start();
-
-        ch.qos.logback.classic.Logger lbLogger = (ch.qos.logback.classic.Logger) LoggerFactory.getLogger(loggerName);
-        lbLogger.addAppender(appender);
-        lbLogger.setLevel(Level.ALL);
+        final org.apache.log4j.Logger log4jLogger = org.apache.log4j.Logger.getLogger(loggerName);
+        log4jLogger.addAppender(appender);
+        log4jLogger.setLevel(Level.ALL);
         return appender;
     }
 
-    public static class TestAppender extends AppenderBase<ILoggingEvent> {
-        final BlockingQueue<ILoggingEvent> queue = new LinkedBlockingQueue<>();
+    private static class TestAppender extends AppenderSkeleton {
+        final BlockingQueue<LoggingEvent> queue = new LinkedBlockingQueue<>();
 
         @Override
-        protected void append(final ILoggingEvent event) {
-            queue.add(event);
+        protected void append(final LoggingEvent loggingEvent) {
+            queue.add(loggingEvent);
+        }
+
+        @Override
+        public void close() throws SecurityException {
+            queue.clear();
+        }
+
+        @Override
+        public boolean requiresLayout() {
+            return false;
         }
     }
 }
